@@ -12,32 +12,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.1.3",
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/core": "^4.0.0",
-    "@angular/forms": "^4.0.0",
-    "@angular/http": "^4.0.0",
-    "@angular/material": "^2.0.0-beta.6",
-    "@angular/cdk": "^2.0.0-beta.6",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@angular/router": "^4.0.0",
-    "@material/layout-grid": "^0.2.0",
-    "@types/lodash": "^4.14.65",
-    "core-js": "^2.4.1",
-    "lodash": "^4.17.4",
-    "moment": "^2.18.1",
-    "ng-http-interceptor": "^3.1.0",
-    "rxjs": "^5.1.0",
-    "videogular2": "^5.4.7",
-    "zone.js": "^0.8.4"
+    "@angular/animations": "^5.2.0",
+    "@angular/cdk": "^5.2.0",
+    "@angular/common": "^5.2.0",
+    "@angular/compiler": "^5.2.0",
+    "@angular/core": "^5.2.0",
+    "@angular/forms": "^5.2.0",
+    "@angular/http": "^5.2.0",
+    "@angular/material": "^5.2.0",
+    "@angular/platform-browser": "^5.2.0",
+    "@angular/platform-browser-dynamic": "^5.2.0",
+    "@angular/router": "^5.2.0",
+    "@types/lodash": "^4.14.107",
+    "core-js": "^2.5.5",
+    "lodash": "^4.17.5",
+    "moment": "^2.22.1",
+    "rxjs": "^5.5.10",
+    "videogular2": "^6.1.1",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "^4.0.0",
-    "@types/core-js": "^0.9.41",
+    "@angular/cli": "^1.7.4",
+    "@angular/compiler-cli": "^5.2.0",
+    "@types/core-js": "^0.9.46",
     "@types/jasmine": "2.5.38",
-    "@types/node": "~6.0.60",
+    "@types/node": "^6.0.106",
     "codelyzer": "~2.0.0",
     "jasmine-core": "~2.5.2",
     "jasmine-spec-reporter": "~3.2.0",
@@ -45,11 +44,11 @@
     "karma-chrome-launcher": "~2.1.1",
     "karma-cli": "~1.0.1",
     "karma-coverage-istanbul-reporter": "^0.2.0",
-    "karma-jasmine": "~1.1.0",
+    "karma-jasmine": "^1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "~5.1.0",
     "ts-node": "~2.0.0",
     "tslint": "~4.5.0",
-    "typescript": "~2.3.0"
+    "typescript": "^2.4.2"
   }
 }
